+++ state.go
@@ -1,0 +1,17 @@
+package xmpp
+
+const (
+	STATE_INIT = iota
+	STATE_FIRST_STREAM
+	STATE_TLS_UPGRADE_REQUESTED
+	STATE_TLS_UPGRADED
+	STATE_TLS_START_STREAM
+	STATE_TLS_AUTH
+	STATE_TLS_REGISTER
+	STATE_AUTHED_START
+	STATE_AUTHED_STREAM
+	STATE_BINDED_STREAM
+	STATE_NORMAL
+	STATE_ERROR
+	STATE_CLOSED
+)
